refactor: contexts
Story: [ch131]
